pyproject.toml
[tool.poetry]
name = "whatsapp-chat-autoexport"
version = "0.1.0"
description = ""
authors = ["AJ Anderson <ajanderson1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
appium-python-client = ">=5.2.4,<6.0.0"
ipykernel = ">=7.1.0,<8.0.0"
colorama = ">=0.4.6,<1.0.0"
google-api-python-client = "^2.100.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.1.0"
tqdm = "^4.66.0"
openai = "^1.0.0"
elevenlabs = "^1.0.0"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.1.0"
pyyaml = "^6.0.3"
rich = "^14.3.1"
typer = "^0.21.1"
textual = ">=0.94.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
pytest-timeout = "^2.1.0"
pytest-asyncio = "^0.23"

[tool.poetry.scripts]
# New unified CLI (recommended)
whatsapp = "whatsapp_chat_autoexport.cli_entry:main"

# Deprecated commands (print migration notice and exit)
whatsapp-export = "whatsapp_chat_autoexport.deprecated_entry:whatsapp_export_main"
whatsapp-process = "whatsapp_chat_autoexport.deprecated_entry:whatsapp_process_main"
whatsapp-drive = "whatsapp_chat_autoexport.deprecated_entry:whatsapp_drive_main"
whatsapp-pipeline = "whatsapp_chat_autoexport.deprecated_entry:whatsapp_pipeline_main"
whatsapp-logs = "whatsapp_chat_autoexport.deprecated_entry:whatsapp_logs_main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=whatsapp_chat_autoexport",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=0",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_api: marks tests that require API keys",
    "requires_device: marks tests that require an Android device",
    "unit: marks tests as unit tests",
]
timeout = 300
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["whatsapp_chat_autoexport"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
